pyproject.toml
[project]
name = "csa-inabox"
version = "0.1.0"
description = "Cloud-Scale Analytics in-a-Box — Azure Data Mesh Platform"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = []

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.0.0",
    "bandit>=1.7.5",
    "pre-commit>=3.5.0",
    # structlog is a governance runtime dep but tests for the shared
    # logging helper need it available in the dev env too.
    "structlog>=24.1.0",
    # Type stubs for third-party libraries that lack inline types — kept
    # here (and not under governance extras) so mypy always sees them
    # during a `pip install -e ".[dev]"` workflow.
    "types-PyYAML>=6.0",
]
governance = [
    "pyyaml>=6.0",
    "azure-monitor-ingestion>=1.0.0",
    # Structured JSON logging with trace/correlation context — see
    # governance/common/logging.py and docs/LOG_SCHEMA.md.
    "structlog>=24.1.0",
    # Great Expectations for the checkpoint runner in
    # governance/dataquality/ge_runner.py. The runner has an in-memory
    # fallback evaluator so this package is not required for the unit
    # tests, but it is required for live Spark-backed checkpoint runs.
    "great-expectations>=1.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["domains", "scripts", "governance", "agent-harness"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "ISC",    # flake8-implicit-str-concat
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "ERA",    # eradicate
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "COM812", # missing trailing comma (conflicts with formatter)
    "ISC001", # single-line implicit string concat (conflicts with formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["domains", "governance"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "--tb=short -q --strict-markers"

[tool.coverage.run]
# Only measure the code under test today.  As new areas grow a real test
# suite (dbt Python runners, Azure Functions, agent harness gates),
# add them to this list alongside their tests.
source = [
    "governance/common",
    "governance/dataquality",
    "governance/contracts",
    "domains/sharedServices/aiEnrichment/functions",
    "domains/sharedServices/eventProcessing/functions",
    "domains/sharedServices/secretRotation/functions",
]
branch = true
omit = [
    "tests/*",
    "**/__init__.py",
    "**/conftest.py",
    "**/test_*.py",
    # The quality-check CLI runner wraps third-party tools (dbt,
    # great_expectations) and is exercised in integration tests, not
    # unit tests.  Keep it out of the coverage gate until an
    # integration test suite exists.
    "governance/dataquality/run_quality_checks.py",
]

[tool.coverage.report]
# CI fails the job if total coverage for the source packages above drops
# below this threshold (see .github/workflows/test.yml).  The threshold
# ratchets upward as the test suite grows — do not lower it without an
# explicit justification in the commit message.
fail_under = 80
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.mypy]
python_version = "3.10"
# Strict mode is enabled globally. New code is expected to pass strict
# checks; when adding a new module that cannot yet comply, prefer adding
# it to the overrides list below with an explicit reason rather than
# weakening this config.
strict = true
# Default target when ``mypy`` is run with no arguments: the governance
# package plus its tests. The two Azure Functions apps both ship a flat
# ``function_app.py`` at different paths, which otherwise collides at
# module resolution time — they are type-checked by dedicated ``mypy
# <path>`` invocations from the Makefile and the test.yml workflow so
# each app resolves as its own module. See ``make typecheck``.
files = [
    "governance",
    "tests",
]
# Third-party Azure SDK libraries still ship partial type info — stay
# lenient there, but keep strict checks on our own code.
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
# Databricks notebooks rely on spark / dbutils / display globals that
# mypy cannot resolve.  Type hints are still added for IDE support, but
# static analysis is disabled so mypy does not fail the CI.
module = [
    "domains.shared.notebooks.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
# Load-test harnesses intentionally subclass dynamically-typed
# third-party base classes (Locust's HttpUser).  They are also not
# part of the deployable toolchain — they run manually via the
# ``load-tests.yml`` workflow_dispatch.  Ignore them so strict mode
# stays useful for the rest of the tree.
module = [
    "tests.load.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
# The Azure Functions host imports function_app.py by path, not as a
# regular Python module, so these files are stand-alone. Test cases do
# not import them as modules either — we only need mypy to check them.
module = [
    "azure.ai.*",
    "azure.cosmos.*",
    "azure.eventhub.*",
]
ignore_missing_imports = true
